+++ app/src/main/java/com/ichi2/apisample/model/DuplicateAddingPrompter.java
@@ -1,4 +1,4 @@
-package com.ichi2.apisample;
+package com.ichi2.apisample.model;
 
 public interface DuplicateAddingPrompter {
     void promptAddDuplicate(MusInterval[] existingMis, DuplicateAddingHandler handler);
