+++ app/src/main/java/com/ichi2/apisample/validation/EmptyValidator.java
@@ -1,0 +1,13 @@
+package com.ichi2.apisample.validation;
+
+public class EmptyValidator implements Validator {
+    @Override
+    public boolean isValid(String value) {
+        return !value.isEmpty();
+    }
+
+    @Override
+    public String getErrorTag() {
+        return "empty";
+    }
+}
