+++ app/src/main/java/com/ichi2/apisample/validation/IntegerRangeValidator.java
@@ -1,0 +1,32 @@
+package com.ichi2.apisample.validation;
+
+public class IntegerRangeValidator implements Validator {
+    private final int minValue;
+    private final int maxValue;
+
+    private String errorTag;
+
+    public IntegerRangeValidator(int minValue, int maxValue) {
+        this.minValue = minValue;
+        this.maxValue = maxValue;
+    }
+
+    @Override
+    public boolean isValid(String value) {
+        int intVal = Integer.parseInt(value);
+        if (intVal < minValue) {
+            errorTag = "below" + minValue;
+            return false;
+        }
+        if (intVal > maxValue) {
+            errorTag = "above" + maxValue;
+            return false;
+        }
+        return true;
+    }
+
+    @Override
+    public String getErrorTag() {
+        return errorTag;
+    }
+}
