+++ app/src/main/java/com/ichi2/apisample/validation/PatternValidator.java
@@ -1,0 +1,19 @@
+package com.ichi2.apisample.validation;
+
+public class PatternValidator implements Validator {
+    private final String pattern;
+
+    public PatternValidator(String pattern) {
+        this.pattern = pattern;
+    }
+
+    @Override
+    public boolean isValid(String value) {
+        return value.matches(pattern);
+    }
+
+    @Override
+    public String getErrorTag() {
+        return "invalid";
+    }
+}
