+++ app/src/main/java/com/ichi2/apisample/validation/Validator.java
@@ -1,0 +1,7 @@
+package com.ichi2.apisample.validation;
+
+public interface Validator {
+    boolean isValid(String value);
+
+    String getErrorTag();
+}
